+++ CHANGES.rst
@@ -3,7 +3,8 @@
 
 Enhancements and Fixes
 ----------------------
-
+- Added functionality to the DatalinkRecordMixin and the DatalinkResultsMixin
+  to handle basic parsing of JSON entries [#709]
 
 
 Deprecations and Removals